requirements.txt
fastapi
uvicorn
jinja2
python-jose[cryptography]
passlib[bcrypt]
pymongo